+++ src/main/java/io/enderdev/endermodpacktweaks/config/CfgTweaks.java
@@ -21,6 +21,11 @@ public class CfgTweaks {
         @Config.Name("[01] Enable Arcane World Tweaks")
         @Config.Comment("This fixes the lag in the other dimensions.")
         public boolean enable = true;
+
+        @Config.RequiresMcRestart
+        @Config.Name("[02] Levitator accepts Redstone")
+        @Config.Comment("Allow disabling the Levitator with a redstone signal.")
+        public boolean redstoneControl = false;
     }
 
     @Config.Name("astral_sorcery")