+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.5.10]
+
+- 🆕 the `Arcane World` Levitator can now be controlled with redstone
+- 🔄 even more performance improvements to the mob health renderer (thanks to tttsaurus)
+- 🔄 removed `modernsplash` from the "incompatible mods list"
+- 🧰 a few small improvements (thanks to rozbrajaczpoziomow)
+
 ## [0.5.9]
 
 - 🆕 set default gamerules via config (setting via world creation screen will be added later)